AboutV1

+++ actions/userModification.php
@@ -2,51 +2,44 @@
 include 'Database.php';
 include 'User.php';
 
-$_SESSION['editShow'] = false;
 $_SESSION['error'] = "";
 $_SESSION['eFormatMail'] = "";
-$database = new Database('blog');
-$query = $database->getHandle()->query("SELECT * FROM user");
-$i = 0;
-foreach ($query as $row){
-    $user = new User($row['user_id'],$row['login'],$row['password'],$row['email'],$row['role']);
-    $_SESSION['users'][$i] = serialize($user);
-    $i++;
-}
-$_SESSION['count'] = $i;
 
-if(!empty($_GET['name'])){
 
-    if($_GET['name'] == "edit"){
-        $_SESSION['editShow'] = true;
-        $_SESSION['idEdit'] = $_GET['id'];
-    }
+if(isset($_POST['action'])) {
 
-    if($_GET['name'] == "delete"){
-        $delete = $database->getHandle()->query('DELETE FROM user WHERE user_id='.$_GET['id'].';');
-        $count = $database->count('user');
-        for($i = $_GET['id']; $i <= $count; $i++)
-       {
-           $update = $database->getHandle()->query('UPDATE user SET user_id="'.$i.'" WHERE user_id="'.($i+1).'"');
-       }
-        header("Location: index.php?action=userModification");
-    }
+    $database = new Database('blog');
+    $query = $database->getHandle()->query("SELECT * FROM user");
 
-    if($_GET['name'] == "editNow"){
-        if(empty($_POST['login']) || empty($_POST['mail'])){
-            $_SESSION['error'] = "Wypełnij dane";
-        }else{
-            if ((filter_var($_POST['mail'], FILTER_VALIDATE_EMAIL))) {
+    if (!empty($_GET['name'])) {
+        if ($_GET['name'] == "delete") {
+            $delete = $database->getHandle()->query('DELETE FROM user WHERE user_id=' . $_GET['id'] . ';');
+            $count = $database->count('user');
+            for ($i = $_GET['id']; $i <= $count; $i++) {
+                $update = $database->getHandle()->query('UPDATE user SET user_id="' . $i . '" WHERE user_id="' . ($i + 1) . '"');
+            }
+            header("Location: index.php?action=about");
+        }
 
-                $idCurrent = $_SESSION['idEdit'];
-                $user = unserialize($_SESSION['users'][$idCurrent - 1]);
-                $update = $database->getHandle()->query('UPDATE user SET user_id="' . $idCurrent . '", login="' . $_POST['login'] . '", password="' . $user->getPassword() . '", email="' . $_POST['mail'] . '", role="' . $_POST['role'] . '"  WHERE user_id="' . $idCurrent . '"');
-                header("Location: index.php?action=userModification");
-            }else{
-                $_SESSION['eFormatMail'] = "Zły format maila";
+        if ($_GET['name'] == "editNow") {
+            if ((empty($_POST['login']) || empty($_POST['mail']))) {
+                $_SESSION['error'] = "Wypełnij dane";
+            } else {
+                if ((filter_var($_POST['mail'], FILTER_VALIDATE_EMAIL))) {
+
+                    $idCurrent = $_SESSION['idEdit'];
+                    $user = unserialize($_SESSION['users'][$idCurrent - 1]);
+                    $update = $database->getHandle()->query('UPDATE user SET user_id="' . $idCurrent . '", login="' . $_POST['login'] . '", password="' . $user->getPassword() . '", email="' . $_POST['mail'] . '", role="' . $_POST['role'] . '"  WHERE user_id="' . $idCurrent . '"');
+
+
+                    header("Location: index.php?action=about");
+                } else {
+                    $_SESSION['eFormatMail'] = "Zły format maila";
+                }
             }
         }
     }
 }
+
